Move python files under src/ directory to follow standard Python conventions. Add basic tests to be run using pytest.
