+++ hikvision-sdk/tests/test_config.py
@@ -1,0 +1,5 @@
+from config import loadConfig
+
+def test_loadConfig():
+    config = loadConfig()
+    assert config is not None
