+++ hikvision-sdk/tests/test_hcnetsdk.py
@@ -1,0 +1,6 @@
+from hcnetsdk import setupSDK
+
+
+# Try to load Hikvision SDK
+def test_setupSDK():
+    setupSDK()
